requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.29.0
certifi==2025.1.31
h11==0.14.0
httpcore==1.0.7
httpx==0.26.0
idna==3.10
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
python-dotenv==1.0.1
python-telegram-bot==20.8
sniffio==1.3.1
SQLAlchemy==2.0.27
typing_extensions==4.12.2
